Route management stuff
More work on route management. The code is in a bad state again and will need to be reformatted, namely PlayerSelection - it should be renamed and the update function should be split into three functions, depending on whether something a hold was hit, a wall was hit or nothing was hit.